#1764 isolated period mode selector
